Refactor: Add WebSocket client tracking and logging
This commit refactors the code to add functionality for tracking WebSocket clients and logging their UUIDs. It introduces a new `clients` array to store the WebSocket clients and adds a hook to assign the `clients` array to the `request` object. Additionally, a log message is added to periodically log the UUIDs of the connected clients.